% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{aligned_read_profile_plots}
\alias{aligned_read_profile_plots}
\title{Returns a plot showing the aligned read profile of HDR and WT reads covering
the highlight_site, and optionally additional plots for top individual deletions.}
\usage{
aligned_read_profile_plots(
  alignment_result,
  num_aligned_read_del_profiles = 3,
  viewing_window = 40
)
}
\arguments{
\item{alignment_result}{Result from a call to alignment_analysis.}

\item{num_aligned_read_del_profiles}{The number of additional plots to show for individual top deletion alleles.}

\item{viewing_window}{Window on either size of the CRISPR cut site to show in the plot.}
}
\value{
Returns a list with a ggplot object for each aligned read profile plot requested (1 for HDR allele + num_aligned_read_del_profiles)
}
\description{
Returns a plot showing the aligned read profile of HDR and WT reads covering
the highlight_site, and optionally additional plots for top individual deletions.
}
\examples{
# Note: First run alignment_analysis()
# ctcf_alignment_results is a pre-loaded result

#aligned_read_profile_plots(ctcf_alignment_results[[1]])
}
\seealso{
\code{\link{alignment_analysis}}
}
