% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{alignment_analysis_plots}
\alias{alignment_analysis_plots}
\title{Returns all main plots for a single alignment analysis result.}
\usage{
alignment_analysis_plots(
  alignment_result,
  opts = genie_plot_options(),
  variance_components_plot = FALSE,
  power_plots = FALSE
)
}
\arguments{
\item{alignment_result}{Result from a call to alignment_analysis.}

\item{opts}{A list with all options needed for rgenie plotting functions.}

\item{variance_components_plot}{If TRUE, then variance_components_plot() is called.}

\item{power_plots}{If TRUE, then power_plots() is called.}
}
\value{
Returns a list of ggplot objects.
}
\description{
Returns all main plots for a single alignment analysis result.
}
\examples{
# Note: First run alignment_analysis()
# mul1_alignment_results is a pre-loaded result

alignment_analysis_plots(mul1_alignment_results[[1]], genie_plot_options())
}
\seealso{
\code{\link{genie_plot_options}}

\code{\link{alignment_analysis}}

\code{\link{alignment_summary_plot}}

\code{\link{deletion_alleles_plot}}

\code{\link{deletion_profile_plot}}

\code{\link{aligned_read_profile_plots}}

\code{\link{replicate_summary_plot}}

\code{\link{replicate_qc_plot}}

\code{\link{allele_effect_plot}}

\code{\link{variance_components_plot}}

\code{\link{power_plots}}
}
