% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{alignment_summary_plot}
\alias{alignment_summary_plot}
\title{Plots a summary of alignment analysis results for a single region.}
\usage{
alignment_summary_plot(alignment_result)
}
\arguments{
\item{alignment_result}{Result from a call to alignment_analysis.}
}
\value{
Returns a ggplot object with a summary of alignment analysis results for a single region.
}
\description{
Plots a summary of alignment analysis results for a single region.
}
\examples{
# Note: First run alignment_analysis()
# mul1_alignment_results is a pre-loaded result

alignment_summary_plot(mul1_alignment_results[[1]])
}
\seealso{
\code{\link{alignment_analysis}}
}
