% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_del_results}
\alias{atac_del_results}
\title{GenIE ATAC example rs7729529 alignment analysis results.}
\format{
A one-item list with an alignment_analysis result object.
}
\usage{
atac_del_results
}
\description{
A list with a single result object from GenIE alignment analysis for the ATAC_rs7729529 example.
}
\keyword{datasets}
