% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_grep_results}
\alias{atac_grep_results}
\title{GenIE ATAC example rs7729529 grep results.}
\format{
A one-item list with a grep_analysis result object.
}
\usage{
atac_grep_results
}
\description{
A list with a single result object from GenIE grep analysis for the ATAC_rs7729529 example.
}
\keyword{datasets}
