% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_replicates}
\alias{atac_replicates}
\title{GenIE replicates for the MUL1 example.}
\format{
A data frame with 12 rows and 5 variables:
\describe{
  \item{name}{region name}
  \item{replicate}{short name of the replicate}
  \item{type}{cDNA" or "gDNA"}
  \item{vp_extraction}{which cDNA/gDNA extraction the replicate is from}
  \item{bam}{path to the BAM file of data for the replicate}
}
}
\usage{
atac_replicates
}
\description{
A data.frame with GenIE replicate details for the MUL1 example.
}
\keyword{datasets}
