% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{deletion_alleles_plot}
\alias{deletion_alleles_plot}
\title{Plots unique deletion alleles and their "pileup" count profile separately for cDNA and gDNA.}
\usage{
deletion_alleles_plot(del_result, viewing_window = 40, color_by = "window")
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis.}

\item{color_by}{A string with one of the values: "none" (default), "window", or "sharing".}
}
\value{
Returns a ggplot object.
}
\description{
Plots unique deletion alleles and their "pileup" count profile separately for cDNA and gDNA.
}
\examples{
del_results = deletion_analysis(regions, replicates)
deletion_alleles_plot(del_results)
}
