% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie.functions.R
\name{deletion_analysis}
\alias{deletion_analysis}
\title{Runs an alignment-based GenIE analysis}
\usage{
deletion_analysis(
  regions,
  replicates,
  required_match_left = 10,
  required_match_right = 10,
  crispr_del_window = 100,
  min_mapq = 0,
  max_mismatch_frac = 0.05,
  min_aligned_bases = 30,
  min_window_overlap = 50,
  exclude_multiple_deletions = F,
  exclude_nonspanning_reads = T,
  ignore_noncrispr_deletions = T,
  allele_profile = F,
  del_span_start = -20,
  del_span_end = 20,
  quiet = FALSE
)
}
\arguments{
\item{regions}{A data.frame defining GenIE regions.}

\item{replicates}{A data.frame defining GenIE replicates.}

\item{required_match_left}{The length of sequence to the left of the HDR site that must exactly match to identify HDR or WT reads.}

\item{required_match_right}{The length of sequence to the right of the HDR site that must exactly match to identify HDR or WT reads.}

\item{crispr_del_window}{The window around the cut site within which any deletion is considered a CRISPR deletion.}

\item{min_mapq}{The minimum mapping quality for reads to be included in the analysis.}

\item{max_mismatch_frac}{The maximum fraction of mismatches a read can have and be included in the analysis.}

\item{min_aligned_bases}{The minimum length of aligned sequence for a read to be included in the analysis.}

\item{exclude_multiple_deletions}{If TRUE, then reads with multiple deletions will be excluded from the analysis.}

\item{ignore_noncrispr_deletions}{If TRUE, then deletions that do not span the region [cut_site - crispr_del_window, cut_site + crispr_del_window] will be ignored; i.e. such reads can be considered HDR or WT.}

\item{allele_profile}{If TRUE, then the result object will contain data.frames named site_profiles and mismatch_profiles, as detailed in the description below.}

\item{del_span_start}{An integer that specifies the start of a window, relative to the region's highlight site, within which deletions are counted.}

\item{del_span_end}{An integer that specifies the end of a window, relative to the region's highlight site, within which deletions are counted.}
}
\value{
Returns a list containing GenIE outputs that depend on the options that were set.
}
\description{
Runs an alignment-based GenIE analysis
}
\examples{
del_results = deletion_analysis(regions, replicates)
}
\seealso{
\code{\link{grep_analysis}}

\code{\link{deletion_plots}}

\code{\link{deletion_summary_plot}}

\code{\link{experiment_summary_plot}}

\code{\link{deletion_alleles_plot}}

\code{\link{deletion_profile_plot}}

\code{\link{replicate_summary_plot}}

\code{\link{replicate_qc_plot}}

\code{\link{allele_effect_plot}}

\code{\link{variance_components_plot}}

\code{\link{power_plots}}
}
