% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie.functions.R
\name{download_example}
\alias{download_example}
\title{Downloads example data for rgenie.}
\usage{
download_example(dir = "~/genie_example", name = "MUL1", quiet = F)
}
\arguments{
\item{dir}{Directory where example data should be put.}
}
\value{
Returns a list containing the same tables as in an individual result,
but concatenated across regions.
}
\description{
The example data is a set of BAM files for GenIE replicates.
}
\examples{
download_example(dir = "~/genie_example", name = "MUL1")
# Data are downloaded and we can run an rgenie analysis
setwd("~/genie_example")
regions = readr::read_tsv("mul1.genie_regions.tsv")
replicates = readr::read_tsv("mul1.genie_replicates.tsv")
grep_results = grep_analysis(regions, replicates)
}
\seealso{
\code{\link{grep_analysis}}

\code{\link{deletion_analysis}}
}
