% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{genie_plot_options}
\alias{genie_plot_options}
\title{Returns a list with default options for all rgenie plots, useful in calls to deletion_plots().}
\usage{
genie_plot_options()
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis and/or grep_analysis.}
}
\value{
Returns a ggplot object.
}
\description{
Returns a list with default options for all rgenie plots, useful in calls to deletion_plots().
}
\examples{
del_result = deletion_analysis(regions, replicates)
opts = genie_plot_options()
plot_list = deletion_plots(del_result, opts)
}
