% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{grep_summary_plot}
\alias{grep_summary_plot}
\title{Plots a summary of genie results from a grep analysis.}
\usage{
grep_summary_plot(grep_result)
}
\arguments{
\item{grep_result}{Result from a call to grep_analysis.}
}
\value{
Returns a ggplot object.
}
\description{
Plots a summary of genie results from a grep analysis.
}
\examples{
# Note: First run grep_analysis()
# mul1_grep_results is a pre-loaded result

grep_summary_plot(mul1_grep_results[[1]])
}
\seealso{
\code{\link{grep_analysis}}
}
