% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mul1_grep_results}
\alias{mul1_grep_results}
\title{GenIE grep results list.}
\format{
A one-item list with a grep_analysis result object.
}
\usage{
mul1_grep_results
}
\description{
A list with a single result object from GenIE grep analysis for the MUL1 example.
}
\keyword{datasets}
