% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mul1_regions}
\alias{mul1_regions}
\title{GenIE regions for the MUL1 example.}
\format{
A data frame with 1 row and 9 variables:
\describe{
  \item{name}{region name}
  \item{sequence_name}{chromosome or amplicon name}
  \item{start}{start coordinate of the region}
  \item{end}{end coordinate of the region}
  \item{highlight_site}{coordinate of the SNP site}
  \item{cut_site}{coordinate of the cut site}
  \item{hdr_allele_profile}{HDR allele profile}
  \item{wt_allele_profile}{WT allele profile}
  \item{ref_sequence}{sequence of the amplicon region}
}
}
\usage{
mul1_regions
}
\description{
A data.frame with GenIE regions details for the MUL1 example.
}
\keyword{datasets}
