% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie_functions.R
\name{power_analysis}
\alias{power_analysis}
\title{Performs estimates of power to detect effects of alleles at different
read fractions, given the variance observed in the alignment_result replicates.}
\usage{
power_analysis(alignment_result, allele_min_reads = 100, WT_fraction = NA)
}
\arguments{
\item{alignment_result}{The result from alignment_analysis}

\item{allele_min_reads}{The minimum number of reads that a deletion allele must have across all replicates to be included.}

\item{WT_fraction}{If specified, then the model will assume this fraction of WT reads}
}
\value{
Returns...
}
\description{
Performs estimates of power to detect effects of alleles at different
read fractions, given the variance observed in the alignment_result replicates.
}
\examples{
# Note: First run alignment_analysis()
# mul1_alignment_results is a pre-loaded result

pwr = power_analysis(mul1_alignment_results[[1]])
power_plots(mul1_alignment_results[[1]])
}
\seealso{
\code{\link{alignment_analysis}}

\code{\link{power_plots}}
}
