% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie.functions.R
\name{replicate_qc_metrics}
\alias{replicate_qc_metrics}
\title{Returns quality control statistics for results from a deletion analysis.}
\usage{
replicate_qc_metrics(
  replicates,
  replicate_alleles,
  max_alleles = 20,
  min_avg_allele_fraction = 0.005,
  exclude_wt = FALSE
)
}
\arguments{
\item{max_alleles}{Param description}

\item{min_avg_allele_fraction}{Param description}

\item{exclude_wt}{Param description}

\item{del_result}{A data.frame defining GenIE regions.}
}
\value{
Returns quality control statistics for results from a deletion analysis.
}
\description{
Returns quality control statistics for results from a deletion analysis.
}
\examples{
del_result = deletion_analysis(regions, replicates)
metrics = replicate_qc_metrics(del_result)
}
\seealso{
\code{\link{deletion_analysis}}

\code{\link{replicate_summary_plot}}

\code{\link{replicate_qc_plot}}
}
