% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{replicate_qc_plot}
\alias{replicate_qc_plot}
\title{Plots quality control metrics for alignment analysis replicates.}
\usage{
replicate_qc_plot(alignment_result, outlier_threshold = NA)
}
\arguments{
\item{alignment_result}{Result from a call to alignment_analysis.}

\item{outlier_threshold}{A numeric threshold for the outlier score, above which replicates will be colored differently.}
}
\value{
Returns a ggplot object with quality control metrics for alignment analysis replicates.
}
\description{
Plots quality control metrics for alignment analysis replicates.
}
\examples{
# Note: First run alignment_analysis()
# mul1_alignment_results is a pre-loaded result

replicate_qc_plot(mul1_alignment_results[[1]])
}
\seealso{
\code{\link{alignment_analysis}}
}
