% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{replicate_summary_plot}
\alias{replicate_summary_plot}
\title{Plots a summary of deletion analysis replicates.}
\usage{
replicate_summary_plot(del_result, outlier_threshold = NA)
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis.}

\item{outlier_threshold}{A numeric threshold for the outlier score, above which replicates will be colored differently.}
}
\value{
Returns a ggplot object with a summary of deletion analysis replicates.
}
\description{
Plots a summary of deletion analysis replicates.
}
\examples{
# Note: First run deletion_analysis()
# mul1_del_results is a pre-loaded result

replicate_summary_plot(mul1_del_results[[1]])
}
\seealso{
\code{\link{deletion_analysis}}
}
