% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie_functions.R
\name{write_results}
\alias{write_results}
\title{Given results for rgenie grep or alignment analysis of a set of regions,
merges together tables across regions.}
\usage{
write_results(results, base_path, delim = "\\t", ext = "")
}
\arguments{
\item{results}{A list of rgenie results from either grep or deletion analyses.}

\item{base_path}{The root path (including any file prefix) to use for saved files. Output files will have names e.g. <base_path>.region_stats.tsv}

\item{delim}{The delimiter to use between columns in saved files.}

\item{ext}{The extension for saved files. If not specified (or an empty string) then the extension is inferred from the delimiter as either "tsv", "csv", or "txt".}
}
\value{
Returns the result of calling bind_results(results).
}
\description{
Given results for rgenie grep or alignment analysis of a set of regions,
merges together tables across regions.
}
\examples{
# Note: First run alignment_analysis()
# mul1_alignment_results is a pre-loaded result

 write_results(mul1_alignment_results, "/path/to/mul1_results")
}
\seealso{
\code{\link{grep_analysis}}

\code{\link{alignment_analysis}}

\code{\link{bind_results}}
}
