% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie.functions.R
\name{bind_results}
\alias{bind_results}
\title{Given results for rgenie grep or deletion analysis of a set of regions,
merges together tables across regions.}
\usage{
bind_results(results)
}
\arguments{
\item{results}{A list of rgenie results from either grep or deletion analyses.}
}
\value{
Returns a list containing the same tables as in an individual result,
but concatenated across regions.
}
\description{
Given results for rgenie grep or deletion analysis of a set of regions,
merges together tables across regions.
}
\examples{
del_results = deletion_analysis(regions, replicates)
del_tables = bind_results(del_results)
}
\seealso{
\code{\link{deletion_analysis}}
}
