% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{deletion_alleles_plot}
\alias{deletion_alleles_plot}
\title{Plots unique deletion alleles and their "pileup" count profile separately for cDNA and gDNA.}
\usage{
deletion_alleles_plot(
  alignment_result,
  viewing_window = 40,
  color_by = "window"
)
}
\arguments{
\item{alignment_result}{Result from a call to alignment_analysis.}

\item{viewing_window}{Window on either size of the CRISPR cut site to show in the plot.}

\item{color_by}{A string with one of the values: "none" (default), "window", or "sharing".}
}
\value{
Returns a ggplot object.
}
\description{
Plots unique deletion alleles and their "pileup" count profile separately for cDNA and gDNA.
}
\examples{
# Note: First run alignment_analysis()
# mul1_alignment_results is a pre-loaded result

deletion_alleles_plot(mul1_alignment_results[[1]])
}
\seealso{
\code{\link{alignment_analysis}}
}
