% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{deletion_plots}
\alias{deletion_plots}
\title{Returns all main plots for a single deletion analysis result.}
\usage{
deletion_plots(
  del_result,
  opts = genie_plot_options(),
  variance_components_plot = F,
  power_plots = F
)
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis.}
}
\value{
Returns a list of ggplot objects.
}
\description{
Returns all main plots for a single deletion analysis result.
}
\examples{
del_result = deletion_analysis(regions, replicates)
deletion_plots(del_result)
}
