% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{deletion_profile_plot}
\alias{deletion_profile_plot}
\title{Plots the deletion "pileup" profile separately for each replicate.}
\usage{
deletion_profile_plot(alignment_result, viewing_window = 40)
}
\arguments{
\item{alignment_result}{Result from a call to alignment_analysis.}

\item{viewing_window}{Window on either size of the CRISPR cut site to show in the plot.}
}
\value{
Returns a ggplot object.
}
\description{
Plots the deletion "pileup" profile separately for each replicate.
}
\examples{
# Note: First run alignment_analysis()
# mul1_alignment_results is a pre-loaded result

deletion_profile_plot(mul1_alignment_results[[1]])
}
\seealso{
\code{\link{alignment_analysis}}
}
