% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{deletion_profile_plot}
\alias{deletion_profile_plot}
\title{Plots the deletion "pileup" profile separately for each replicate.}
\usage{
deletion_profile_plot(del_result, viewing_window = 40)
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis.}

\item{viewing_window}{Window on either size of the CRISPR cut site to show in the plot.}
}
\value{
Returns a ggplot object.
}
\description{
Plots the deletion "pileup" profile separately for each replicate.
}
\examples{
\dontrun{
del_results = deletion_analysis(regions, replicates)
deletion_profile_plot(del_results)
}
}
\seealso{
\code{\link{deletion_analysis}}
}
