% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{experiment_summary_plot}
\alias{experiment_summary_plot}
\title{Plots a summary of genie results across multiple regions.}
\usage{
experiment_summary_plot(grep_results, alignment_results)
}
\arguments{
\item{grep_results}{Result from a call to grep_analysis (or NULL).}

\item{alignment_results}{Result from a call to alignment_analysis (or NULL).}
}
\value{
Returns a ggplot object.
}
\description{
Plots a summary of genie results across multiple regions.
}
\examples{
# Note: First run alignment_analysis() and/or grep_analysis()
# mul1_alignment_results and mul1_grep_results are pre-loaded

experiment_summary_plot(mul1_grep_results, mul1_alignment_results)
}
\seealso{
\code{\link{grep_analysis}}

\code{\link{alignment_analysis}}

\code{\link{bind_results}}
}
