% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{experiment_summary_plot}
\alias{experiment_summary_plot}
\title{Plots a summary of genie results across multiple regions.}
\usage{
experiment_summary_plot(grep_results, del_results)
}
\arguments{
\item{grep_results}{Result from a call to grep_analysis (or NULL).}

\item{del_results}{Result from a call to deletion_analysis (or NULL).}
}
\value{
Returns a ggplot object.
}
\description{
Plots a summary of genie results across multiple regions.
}
\examples{
\dontrun{
grep_results = grep_analysis(regions, replicates)
del_results = deletion_analysis(regions, replicates)
experiment_summary_plot(grep_results, del_results)
}
}
\seealso{
\code{\link{grep_analysis}}

\code{\link{deletion_analysis}}

\code{\link{bind_results}}
}
