% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{experiment_summary_plot}
\alias{experiment_summary_plot}
\title{Plots a summary of genie results across multiple regions.}
\usage{
experiment_summary_plot(grep_results, del_results)
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis and/or grep_analysis.}
}
\value{
Returns a ggplot object.
}
\description{
Plots a summary of genie results across multiple regions.
}
\examples{
grep_results = grep_analysis(regions, replicates)
del_results = deletion_analysis(regions, replicates)
experiment_summary_plot(grep_results, del_results)
}
