% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{genie_plot_options}
\alias{genie_plot_options}
\title{Returns a list with default options for all rgenie plots, useful in calls to deletion_plots().}
\usage{
genie_plot_options()
}
\description{
Returns a list with default options for all rgenie plots, useful in calls to deletion_plots().
}
\examples{
\dontrun{
del_result = deletion_analysis(regions, replicates)
opts = genie_plot_options()
plot_list = deletion_plots(del_result, opts)
}
}
\seealso{
\code{\link{deletion_analysis}}
}
