% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mul1_del_results}
\alias{mul1_del_results}
\title{GenIE alignment analysis results for the MUL1 example.}
\format{
A one-item list with a alignment_analysis result object.
}
\usage{
mul1_del_results
}
\description{
A list with a single result object from GenIE alignment analysis for the MUL1 example.
}
\keyword{datasets}
