% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{power_plots}
\alias{power_plots}
\title{Returns a set of plots summarising the expected power to detect a significant
effect given various allele fraction and effect size combinations.}
\usage{
power_plots(del_result, allele_min_reads = 100, WT_fraction = NA)
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis.}

\item{viewing_window}{Window on either size of the CRISPR cut site to show in the plot.}
}
\value{
Returns a ggplot object.
}
\description{
Returns a set of plots summarising the expected power to detect a significant
effect given various allele fraction and effect size combinations.
}
\examples{
del_results = deletion_analysis(regions, replicates)
power_plots(del_results)
}
