% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{replicate_qc_plot}
\alias{replicate_qc_plot}
\title{Plots quality control metrics for deletion analysis replicates.}
\usage{
replicate_qc_plot(del_result, outlier_threshold = NA)
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis.}
}
\value{
Returns a ggplot object with quality control metrics for deletion analysis replicates.
}
\description{
Plots quality control metrics for deletion analysis replicates.
}
\examples{
del_results = deletion_analysis(regions, replicates)
replicate_qc_plot(del_results)
}
