% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{replicate_summary_plot}
\alias{replicate_summary_plot}
\title{Plots a summary of deletion analysis replicates.}
\usage{
replicate_summary_plot(del_result, outlier_threshold = NA)
}
\arguments{
\item{del_result}{Result from a call to deletion_analysis.}
}
\value{
Returns a ggplot object with a summary of deletion analysis replicates.
}
\description{
Plots a summary of deletion analysis replicates.
}
\examples{
del_results = deletion_analysis(regions, replicates)
replicate_summary_plot(del_results)
}
