% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{variance_components_plot}
\alias{variance_components_plot}
\title{Plots variance components estimates for all unique alleles.}
\usage{
variance_components_plot(varcomp, split_by_fraction = F)
}
\arguments{
\item{varcomp}{Result from a call to get_variance_components.}

\item{split_by_fraction}{If TRUE, then points are colored by allele fraction.}
}
\value{
Returns a ggplot object.
}
\description{
Plots variance components estimates for all unique alleles.
}
\examples{
\dontrun{
del_results = deletion_analysis(regions, replicates)
vc = get_variance_components(del_result[[1]], replicates)
variance_components_plot(vc)
}
}
\seealso{
\code{\link{deletion_analysis}}

\code{\link{get_variance_components}}
}
