% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{variance_components_plot}
\alias{variance_components_plot}
\title{Plots variance components estimates for all unique alleles.}
\usage{
variance_components_plot(varcomp, split_by_fraction = F)
}
\arguments{
\item{split_by_fraction}{If TRUE, then points are colored by allele fraction.}

\item{del_result}{Result from a call to deletion_analysis.}
}
\value{
Returns a ggplot object.
}
\description{
Plots variance components estimates for all unique alleles.
}
\examples{
del_results = deletion_analysis(regions, replicates)
variance_components_plot(del_results)
}
